// src/types/tailwind.types.ts
export type TButtonVariant =
	| 'primary'
	| 'secondary'
	| 'ghost'
	| 'outline'
	| 'accent'
	| 'warning'
	| 'success'
	| 'error'
	| 'outline'
	| 'disabled'
export type TInputVariant =
	| 'bordered'
	| 'ghost'
	| 'secondary'
	| 'accent'
	| 'info'
	| 'success'
	| 'warning'
	| 'error'

export type TSize = 'xs' | 'sm' | 'md' | 'lg'
export type TDirection = 'end' | 'top' | 'bottom' | 'left' | 'right'
export type TRadius = 'none' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl' | 'full'
export type TShadow = 'sm' | 'md' | 'lg' | 'xl' | '2xl' | 'none' | 'shadow'